+++ frontend/package.json
@@ -4,7 +4,7 @@
   "description": "OKCandidate frontend app",
   "main": "index.js",
   "scripts": {
-    "dev": "next",
+    "dev": "NODE_ENV=development node index.js",
     "build": "next build",
     "start": "next start"
   },
@@ -12,12 +12,14 @@
   "license": "MIT",
   "dependencies": {
     "@zeit/next-sass": "^0.1.1",
+    "express": "^4.16.3",
     "next": "^5.0.0",
     "next-redux-wrapper": "^1.3.5",
     "node-sass": "^4.8.0",
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-rating": "^0.7.1",
     "react-redux": "^5.0.7",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
