+++ frontend/pages/index.js
@@ -12,7 +12,7 @@ export default () => (
                     <h1>Don't go to the polls alone.</h1>
                     <p>Take your OKCandidate</p>
                 </div>
-                <a href="/start" className="button large primary">Get Started</a>
+                <a href="/surveys" className="button large primary">Get Started</a>
             </div>
 
             <div className="container">
